requirements.txt
Flask
absl-py==0.12.0
astroid==2.5.3
backcall==0.2.0
cachetools==4.2.1
certifi==2020.12.5
chardet==4.0.0
cycler==0.10.0
Cython==0.29.23
decorator==5.0.5
google-auth==1.28.1
google-auth-oauthlib==0.4.4
grpcio==1.37.0
idna==2.10
ipython==7.22.0
ipython-genutils==0.2.0
isort==5.8.0
jedi==0.18.0
kiwisolver==1.3.1
lazy-object-proxy==1.6.0
Markdown==3.3.4
matplotlib==3.4.1
mccabe==0.6.1
numpy==1.20.2
oauthlib==3.1.0
opencv-python==4.5.1.48
pandas==1.2.4
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.2.0
prompt-toolkit==3.0.18
protobuf==3.15.8
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.8.1
pylint==2.7.4
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
scipy==1.6.2
seaborn==0.11.1
six==1.15.0
tensorboard==2.4.1
tensorboard-plugin-wit==1.8.0
thop==0.0.31.post2005241907
toml==0.10.2
torch==1.8.1
torchaudio==0.8.1
torchvision==0.9.1
tqdm==4.60.0
traitlets==5.0.5
typing-extensions==3.7.4.3
urllib3==1.26.4
watchdog==2.0.3
wcwidth==0.2.5
Werkzeug==1.0.1
wrapt==1.12.1
